refactor RenterList
